@@ -1,17 +1,16 @@
+using System;
 using Microsoft.Maui;
 using Microsoft.Maui.Hosting;
-using System;
 
-namespace TipCalculator
+namespace TipCalculator;
+
+class Program : MauiApplication
 {
-    internal class Program : MauiApplication
-    {
-        protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
+	protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
 
-        static void Main(string[] args)
-        {
-            var app = new Program();
-            app.Run(args);
-        }
-    }
+	static void Main(string[] args)
+	{
+		var app = new Program();
+		app.Run(args);
+	}
 }
